--- app/components/vertical_ellipsis_menu/component.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module VerticalEllipsisMenu
-  class Component < ViewComponent::Base
-  end
-end
