+++ app/components/vertical_ellipsis_menu_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class VerticalEllipsisMenuComponent < ViewComponent::Base
+end
