+++ app/components/vertical_ellipsis_menu_component/vertical_ellipsis_menu_component.scss
