+++ app/webpacker/css/admin_v3/all.scss
@@ -127,7 +127,7 @@
 @import "components/tom_select"; // admin_v3
 
 @import "app/components/modal_component/modal_component";
-@import "app/components/vertical_ellipsis_menu/component"; // admin_v3 and only V3
+@import "app/components/vertical_ellipsis_menu_component/vertical_ellipsis_menu_component"; // admin_v3 and only V3
 @import "app/components/tag_list_input_component/tag_list_input_component";
 @import "app/webpacker/css/admin/trix.scss";
